requirements.txt
asgiref==3.9.2
attrs==25.3.0
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
Faker==37.8.0
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.6.4
packaging==25.0
pillow==11.3.0
pip-tools==7.5.0
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
PySocks==1.7.1
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
